ya se pueden borrar todas las dimensiones de un hijo, se pueden borrar dimensiones con recuerdos, se pueden borrar hijos con dimensiones y recuerdos

+++ Ubefirstapp/HijosConfigViewController.swift
@@ -52,47 +52,64 @@ class HijosConfigViewController: UIViewController, UITableViewDataSource, UITabl
     @IBOutlet weak var btn_eliminar: UIButton!
     
     @IBAction func btnPressed_eliminar(_ sender: Any) {
+        let database = LoadDataToFirestore()
         let alertController = UIAlertController(title: "¿Eliminar hijo " + userData.hijos[lastHijoConfigPersonTap].nombre + " ?", message: "Esta acción eliminara permanentemente toda su información incluyendo dimensiones y recuerdos, para continuar introduzca su nombre de usuario", preferredStyle: UIAlertController.Style.alert)
         alertController.addTextField(configurationHandler: self.usernameTextField)
         alertController.addAction(UIAlertAction(title: "Borrar", style: UIAlertAction.Style.destructive) { UIAlertAction in
+            
             if (self.usernameTextField?.text == userData.nombre){
                 let alertController2 = UIAlertController(title: "Borrando a " + userData.hijos[lastHijoConfigPersonTap].nombre , message: "Espere un momento porfavor", preferredStyle: UIAlertController.Style.alert)
                 self.present(alertController2, animated: true, completion: nil)
-                db.collection("users").document(userData.uid).collection("hijos").document(userData.hijosref[lastHijoConfigPersonTap]).collection("dimensiones").getDocuments(){ (querySnapshot, err) in
+                
+                //obtengo todas las dimensiones del hijo
+                db.collection("users").document(userData.uid).collection("hijos").document(userData.hijosref[lastHijoConfigPersonTap]).collection("dimensiones").getDocuments(){ (querySnapshot1, err) in
                     if err != nil {
-                        print(err!)
+                        print(err)
                     } else {
-                        var i = 0
-                        
-                        for document in querySnapshot!.documents {
-                            print(document.documentID)
-                            db.collection("users").document(userData.uid).collection("hijos").document(userData.hijosref[lastHijoConfigPersonTap]).collection("dimensiones").document(document.documentID).delete(){ err in
-                                if let err = err {
-                                    print("Error removing document: \(err)")
+                        for document1 in querySnapshot1!.documents {
+                            //elimino cada recuerdo de cada dimension
+                            db.collection("users").document(userData.uid).collection("hijos").document(userData.hijosref[lastHijoConfigPersonTap]).collection("dimensiones").document(document1.documentID).collection("recuerdos").getDocuments(){ (querySnapshot2, err) in
+                                if err != nil {
+                                    print(err)
                                 } else {
-                                    i = i + 1
-                                    print("Document successfully removed!")
-                                    if (i == userData.hijos[lastHijoConfigPersonTap].dimensiones.count){
-                                        db.collection("users").document(userData.uid).collection("hijos").document(userData.hijosref[lastHijoConfigPersonTap]).delete() { err in
+                                    for document2 in querySnapshot2!.documents{
+                                        //si la dimension tiene recuerdos, aqui se elimina cada uno de los recuerdos
+                                        db.collection("users").document(userData.uid).collection("hijos").document(userData.hijosref[lastHijoConfigPersonTap]).collection("dimensiones").document(document1.documentID).collection("recuerdos").document(document2.documentID).delete(){ err in
                                             if let err = err {
                                                 print("Error removing document: \(err)")
                                             } else {
                                                 print("Document successfully removed!")
-                                                alertController2.dismiss(animated: true, completion: nil)
-                                                userData.hijos.remove(at: lastHijoConfigPersonTap)
-                                                userData.hijosref.remove(at: lastHijoConfigPersonTap)
-                                                let loadDataBase = LoadDataToFirestore()
-                                                loadDataBase.eliminarReferenciaHijo()
-                                                self.navigationController?.popViewController(animated: true)
-                                                
                                             }
                                         }
                                     }
                                 }
+                                    
+                                    //si no existen recuerdos en el snapshot entonces borro la dimension, aqui entra cuando se eliminar todos los recuerdos o cuando no tiene recuerdos la dimension
+                                    db.collection("users").document(userData.uid).collection("hijos").document(userData.hijosref[lastHijoConfigPersonTap]).collection("dimensiones").document(document1.documentID).delete() { err in
+                                        if let err = err {
+                                            print("Error removing document: \(err)")
+                                        } else {
+                                            print("Document successfully removed!")
+                                        }
+                                    }
+                                }
+                            }
+                            //aqui se elimina al hijo
+                            db.collection("users").document(userData.uid).collection("hijos").document(userData.hijosref[lastHijoConfigPersonTap]).delete(){ err in
+                                if let err = err {
+                                    print("Error removing document: \(err)")
+                                } else {
+                                    print("Document successfully removed!")
+                                    userData.hijos.remove(at: lastHijoConfigPersonTap)
+                                    userData.hijosref.remove(at: lastHijoConfigPersonTap)
+                                    database.eliminarReferenciaHijo()
+                                    alertController2.dismiss(animated: true, completion: nil)
+                                    self.navigationController?.popViewController(animated: true)
+                                }
                             }
                         }
                     }
-                }
+                
             }else{
                 self.displayUsernameIncorrect()
             }
@@ -162,10 +179,20 @@ class HijosConfigViewController: UIViewController, UITableViewDataSource, UITabl
     }
     
     func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath){
-        let alert = UIAlertController(title: "¿Eliminar dimensión " + userData.hijos[lastHijoConfigPersonTap].dimensiones[indexPath.row].nombre + "?", message: "Está acción borrara permanentemente la dimensión y todos sus recuerdos contenidos", preferredStyle: UIAlertController.Style.alert)
+        var title = ""
+        var subtitle = ""
+        if (userData.hijos[lastHijoConfigPersonTap].dimensiones.count == 1){
+            title = "Hijo sin dimensiones"
+            subtitle = "El hijo " + userData.hijos[lastHijoConfigPersonTap].nombre + " solo tiene la dimensión " + userData.hijos[lastHijoConfigPersonTap].dimensiones[indexPath.row].nombre + ", al eliminarla igualmente se eliminara toda la información del hijo. Introduzca su nombre de usuario para continuar"
+        }else{
+            title = "¿Eliminar dimensión " + userData.hijos[lastHijoConfigPersonTap].dimensiones[indexPath.row].nombre + "?"
+            subtitle = "Está acción borrara permanentemente la dimensión y todos sus recuerdos contenidos"
+        }
+        
+        let alert = UIAlertController(title: title, message: subtitle, preferredStyle: UIAlertController.Style.alert)
         alert.addTextField(configurationHandler: self.usernameTextField)
 
-        alert.addAction(UIAlertAction(title: "Yes", style: UIAlertAction.Style.destructive, handler: {(ACTION) in alert.dismiss(animated: true, completion: nil)
+        alert.addAction(UIAlertAction(title: "Eliminar", style: UIAlertAction.Style.destructive, handler: {(ACTION) in alert.dismiss(animated: true, completion: nil)
             
             if editingStyle == .delete{
                 if (userData.nombre == self.usernameTextField?.text){
@@ -183,7 +210,7 @@ class HijosConfigViewController: UIViewController, UITableViewDataSource, UITabl
             }
         }))
         
-        alert.addAction(UIAlertAction(title: "No", style: UIAlertAction.Style.cancel, handler: { (ACTION) in alert.dismiss(animated: true, completion: nil)
+        alert.addAction(UIAlertAction(title: "Cancelar", style: UIAlertAction.Style.cancel, handler: { (ACTION) in alert.dismiss(animated: true, completion: nil)
         }))
         
         self.present(alert, animated: true, completion: nil)
@@ -191,7 +218,7 @@ class HijosConfigViewController: UIViewController, UITableViewDataSource, UITabl
     
     func deleteDimension(hijoID: String, dimensionID: String, dimensionIndex: Int, hijoIndex:Int, RecuerdosCount:Int, indexPath: IndexPath){
         let database = LoadDataToFirestore()
-        let alertController4 = UIAlertController(title: "Borrando dimensión " + userData.hijos[lastHijoConfigPersonTap].dimensiones[indexPath.row].nombre , message: "Espere un momento porfavor", preferredStyle: UIAlertController.Style.alert)
+        let alertController4 = UIAlertController(title: "Borrando dimensión " + userData.hijos[lastHijoConfigPersonTap].dimensiones[indexPath.row].nombre, message: "Espere un momento porfavor", preferredStyle: UIAlertController.Style.alert)
         self.present(alertController4, animated: true, completion: nil)
         
         //obtengo snapshot de todos los documentos de esa coleccion
@@ -201,18 +228,23 @@ class HijosConfigViewController: UIViewController, UITableViewDataSource, UITabl
             } else {
                 var i = 0
                 if (querySnapshot?.documents.count as! Int == 0){
-                    db.collection("users").document(userData.uid).collection("hijos").document(hijoID).collection("dimensiones").document(dimensionID).delete() { err in
-                        if let err = err {
-                            print("Error removing document: \(err)")
-                        } else {
-                            print("Document successfully removed!")
-                            userData.hijos[hijoIndex].dimensionesref.remove(at: indexPath.row)
-                            userData.hijos[lastHijoConfigPersonTap].dimensiones.remove(at: indexPath.row)
-                            database.deleteReferenceDimension(hijoIndex:hijoIndex)
-                            self.tableView_dimensiones.beginUpdates()
-                            self.tableView_dimensiones.deleteRows(at: [indexPath], with: .fade)
-                            self.tableView_dimensiones.endUpdates()
-                            alertController4.dismiss(animated: true, completion: nil)
+                    if (userData.hijos[lastHijoConfigPersonTap].dimensiones.count == 1){
+                        //alertController4.dismiss(animated: true, completion: nil)
+                        self.lastDimensionAction(indexPath: indexPath, alertController: alertController4)
+                    }else{
+                        db.collection("users").document(userData.uid).collection("hijos").document(hijoID).collection("dimensiones").document(dimensionID).delete() { err in
+                            if let err = err {
+                                print("Error removing document: \(err)")
+                            } else {
+                                print("Document successfully removed!")
+                                userData.hijos[hijoIndex].dimensionesref.remove(at: indexPath.row)
+                                userData.hijos[lastHijoConfigPersonTap].dimensiones.remove(at: indexPath.row)
+                                database.deleteReferenceDimension(hijoIndex:hijoIndex)
+                                self.tableView_dimensiones.beginUpdates()
+                                self.tableView_dimensiones.deleteRows(at: [indexPath], with: .fade)
+                                self.tableView_dimensiones.endUpdates()
+                                alertController4.dismiss(animated: true, completion: nil)
+                            }
                         }
                     }
                     
@@ -228,21 +260,23 @@ class HijosConfigViewController: UIViewController, UITableViewDataSource, UITabl
                             print("Document successfully removed!")
                             if (i == RecuerdosCount){
                                 //elimino el documento que contiene a la coleccion borrada, es decir el doc con el nombre de la dimension
-                                db.collection("users").document(userData.uid).collection("hijos").document(hijoID).collection("dimensiones").document(dimensionID).delete() { err in
-                                    if let err = err {
-                                        print("Error removing document: \(err)")
-                                    } else {
-                                        print("Document successfully removed!")
-                                        userData.hijos[hijoIndex].dimensionesref.remove(at: indexPath.row)
-                                        userData.hijos[lastHijoConfigPersonTap].dimensiones.remove(at: indexPath.row)
-                                        database.deleteReferenceDimension(hijoIndex:hijoIndex)
-                                        self.tableView_dimensiones.beginUpdates()
-                                        self.tableView_dimensiones.deleteRows(at: [indexPath], with: .fade)
-                                        self.tableView_dimensiones.endUpdates()
-                                        alertController4.dismiss(animated: true, completion: nil)
-                                        //alertController2.dismiss(animated: true, completion: nil)
-                                        //self.navigationController?.popViewController(animated: true)
-                                        
+                                if (userData.hijos[lastHijoConfigPersonTap].dimensiones.count == 1){
+                                    alertController4.dismiss(animated: true, completion: nil)
+                                    self.lastDimensionAction(indexPath: indexPath, alertController: alertController4)
+                                }else{
+                                    db.collection("users").document(userData.uid).collection("hijos").document(hijoID).collection("dimensiones").document(dimensionID).delete() { err in
+                                        if let err = err {
+                                            print("Error removing document: \(err)")
+                                        } else {
+                                            print("Document successfully removed!")
+                                            userData.hijos[hijoIndex].dimensionesref.remove(at: indexPath.row)
+                                            userData.hijos[lastHijoConfigPersonTap].dimensiones.remove(at: indexPath.row)
+                                            database.deleteReferenceDimension(hijoIndex:hijoIndex)
+                                            self.tableView_dimensiones.beginUpdates()
+                                            self.tableView_dimensiones.deleteRows(at: [indexPath], with: .fade)
+                                            self.tableView_dimensiones.endUpdates()
+                                            alertController4.dismiss(animated: true, completion: nil)
+                                        }
                                     }
                                 }
                             }
@@ -253,6 +287,35 @@ class HijosConfigViewController: UIViewController, UITableViewDataSource, UITabl
         }
         
     }
+    
+    func lastDimensionAction(indexPath: IndexPath, alertController: UIAlertController) {
+        let dbtemp = LoadDataToFirestore()
+        db.collection("users").document(userData.uid).collection("hijos").document(userData.hijosref[lastHijoConfigPersonTap]).collection("dimensiones").document(userData.hijos[lastHijoConfigPersonTap].dimensionesref[indexPath.row]).delete() { err in
+            if let err = err {
+                print("Error removing document: \(err)")
+            } else {
+                print("Document successfully removed!")
+                db.collection("users").document(userData.uid).collection("hijos").document(userData.hijosref[lastHijoConfigPersonTap]).delete() { err in
+                    if let err = err {
+                        print("Error removing document: \(err)")
+                    } else {
+                        print("Document successfully removed!")
+                        userData.hijos.remove(at: lastHijoConfigPersonTap)
+                        userData.hijosref.remove(at: lastHijoConfigPersonTap)
+                        dbtemp.eliminarReferenciaHijo()
+                        /*self.tableView_dimensiones.beginUpdates()
+                        self.tableView_dimensiones.deleteRows(at: [indexPath], with: .fade)
+                        self.tableView_dimensiones.endUpdates()*/
+                        alertController.dismiss(animated: true, completion: nil)
+                        self.navigationController?.popViewController(animated: true)
+                        
+                    }
+                }
+            }
+        }
+    }
+    
+    
     
     /*
     // MARK: - Navigation
